requirements/base.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.51.0
django-model-utils==4.3.1
django-notifications-hq==1.7.0
django-simplemde==0.1.4
django-social-share==2.3.0
django-taggit==3.1.0
djangorestframework==3.14.0
idna==3.4
jsonfield==3.1.0
markdown2==2.4.7
oauthlib==3.2.2
pdfkit==1.0.0
Pillow==9.4.0
psycopg2==2.9.5
pycld2==0.41
pycparser==2.21
PyJWT==2.6.0
pymemcache==4.0.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
swapper==1.3.0
urllib3==1.26.14